backend/requirements-dev.txt
# SSL Monitor Pro - Development Dependencies
# Additional packages needed for development, testing, and debugging

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
pytest-html==4.1.1

# Code Quality and Linting
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0

# Development Tools
ipython==8.17.2
jupyter==1.0.0
notebook==7.0.6

# API Documentation and Testing
httpx==0.25.2
requests==2.31.0
responses==0.24.1

# Database Testing
pytest-postgresql==5.0.0
factory-boy==3.3.0
faker==20.1.0

# Performance and Profiling
memory-profiler==0.61.0
py-spy==0.3.14
line-profiler==4.1.1

# Security Testing
bandit==1.7.5
safety==2.3.5

# Development Utilities
python-dotenv==1.0.0
rich==13.7.0
click==8.1.7

# Hot Reloading and Development Server
watchfiles==0.21.0
uvloop==0.19.0

# Debugging
pdb++==0.10.3
ipdb==0.13.13

# Mock and Fixtures
freezegun==1.2.2
pytest-fixture-order==0.1.1

# Coverage and Reporting
coverage==7.3.2
coveralls==3.3.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
myst-parser==2.0.0

# Load Testing
locust==2.17.0

# Database Migrations and Management
alembic==1.12.1
sqlalchemy-utils==0.41.1

# Environment and Configuration
python-decouple==3.8
pydantic-settings==2.1.0

# Async Testing
aioresponses==0.7.4
asyncio-mock==0.1.1

# Time and Date Testing
freezegun==1.2.2

# SSL and Certificate Testing
cryptography==41.0.7
pyOpenSSL==23.3.0

# Email Testing
fakeredis==2.20.1

# Monitoring and Metrics (Development)
prometheus-client==0.19.0

# Additional FastAPI Development Tools
fastapi-cli==0.0.3
